third_party/WebKit/LayoutTests/css3/filters/usecounter-filter-function-no-arguments.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style></style>
<script>
'use strict';

test(() => {
    const CSSFilterFunctionNoArguments = 1403; // From UseCounter.h
    var style = document.querySelector('style');
    assert_false(internals.isUseCounted(document, CSSFilterFunctionNoArguments));
    style.textContent = "* { filter: brightness(); }";
    assert_true(internals.isUseCounted(document, CSSFilterFunctionNoArguments));
}, "'filter' function without arguments is use counted.");
</script>
